Adding 'i3' configurations files
This includes the use of 'conky' as status bar, some scripts to deal with
the environment, 'dunst' for notifications (through libnotify), and
'rofi' as an application launcher.

Special attention (for now) on the 'amixer-wrapper', that control some
basic amixer functions (vol +/-, mute) and also gets some information
from the sound device (right now just returning if it's muted, or its
current volume level).

It's the minimal set of settings for me to start using i3 without other
desktop environments (I like KDE, and will still using some of its
softwares, like dolphin and KDEConnect), but probably will do a fine
tunning based on my use of the window manager.

So, finally, it's all solarized! Slightly.

Signed-off-by: Rodrigo Oliveira <rodrigo@deadbodyoutline.com>